+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.12.0] - 2019-12-04
 - Added `Env::make_vector` and `Env::vector` functions.
 - Added a default implementation for `Transfer::type_name`. This raised the minimum supported Rust version to 1.38 (for `std::any::type_name()`).
 - Added function to copy a Lisp string's content to a buffer `Value::copy_string_contents`.
@@ -97,7 +99,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.2.0] - 2018-01-04
 New reworked version
 
-[Unreleased]: https://github.com/ubolonton/emacs-module-rs/compare/0.11.0...HEAD
+[Unreleased]: https://github.com/ubolonton/emacs-module-rs/compare/0.12.0...HEAD
+[0.12.0]: https://github.com/ubolonton/emacs-module-rs/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.3...0.11.0
 [0.10.3]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.2...0.10.3
 [0.10.2]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.1...0.10.2